feat: add sanitization
