+++ server/routes/controller/wishcard.controller.js
@@ -1,5 +1,6 @@
 const moment = require('moment');
 const WishCardRepository = require('../../db/repository/WishCardRepository');
+const UserRepository = require('../../db/repository/UserRepository');
 
 async function getWishCardSearchResult(
   itemName,
@@ -33,6 +34,30 @@ async function getWishCardSearchResult(
   return allWishCards.filter((item) => (childAge < 15 ? item.age < childAge : item.age > childAge));
 }
 
+async function getLockedWishCards(req) {
+  const response = {
+    wishCardId: req.params.id,
+  };
+
+  if (!req.session.user) {
+    response.error = 'User not found';
+    return response;
+  }
+
+  const user = await UserRepository.getUserByObjectId(req.session.user._id);
+  if (!user) {
+    response.error = 'User not found';
+    return response;
+  }
+  response.userId = user._id;
+  response.alreadyLockedWishCard = await WishCardRepository.getLockedWishcardsByUserId(
+    req.session.user._id,
+  );
+
+  return response;
+}
+
 module.exports = {
+  getLockedWishCards,
   getWishCardSearchResult,
 };
